LayoutTests/js/dom/modules/import-from-set-interval.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test import from setInterval.');

// Module will be executed asynchronously.
window.jsTestIsAsync = true;
window.passedCocoa = null;
window.id = null;

function loaded(cocoa)
{
    window.passedCocoa = cocoa;
    shouldBeTrue(`window.imported`);
    shouldBe(`window.passedCocoa`, `window.exportedCocoa`);
    finishJSTest();
}
</script>
<script>
</script>
<script src="../../../resources/js-test-post.js"></script>
<script type="module">
window.id = setInterval(`clearInterval(window.id);import("./resources/import-src-simple.js").then((ns) => loaded(ns.ok));`, 0);
</script>
</body>
</html>
